modify

+++ proj1b/OffByN.java
@@ -12,6 +12,6 @@ public class OffByN implements CharacterComparator  {
 
     @Override
     public boolean equalChars(char x, char y) {
-        return Math.abs(x - y) == n;
+        return Math.abs((int)x - (int)y) == n;
     }
 }
